--- a/google/cloud/internal/base64_transforms.h
+++ b/google/cloud/internal/base64_transforms.h
@@ -94,6 +94,26 @@ struct Base64Decoder {
   std::string const& rep_;  // encoded
 };
 
+Status Base64DecodingError(std::string const& base64,
+                           std::string::const_iterator p);
+
+std::pair<std::array<unsigned char, 3>, std::size_t> Base64Fill(
+    unsigned char p0, unsigned char p1, unsigned char p2, unsigned char p3);
+
+template <typename Sink>
+Status FromBase64(std::string const& base64, Sink const& sink) {
+  auto p = base64.begin();
+  for (; std::distance(p, base64.end()) >= 4; p = std::next(p, 4)) {
+    auto r = Base64Fill(*p, *(p + 1), *(p + 2), *(p + 3));
+    if (r.second == 0) return Base64DecodingError(base64, p);
+    for (auto* o = r.first.begin(); o != r.first.begin() + r.second; ++o) {
+      sink(*o);
+    }
+  }
+  if (p != base64.end()) return Base64DecodingError(base64, p);
+  return Status{};
+}
+
 Status ValidateBase64String(std::string const& input);
 
 }  // namespace internal