Use the transforms in `google::cloud::internal::` to convert from and to
base64 encoding. For the moment these continue to through exceptions on
errors, a future PR will convert the remaining exception to
`StatusOr<>`.

I performed some refactoring in the base64 transforms: in spanner the
base64 decoding is always performed after the source is validated, in
storage that is not the case.

Part of the work for #6946

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/googleapis/google-cloud-cpp/6956)
<!-- Reviewable:end -->