@coryan 2021-07-09
The changes are within the margin of error performance-wise:

## Before (v1.29.x)

```
Run on (96 X 2000.16 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 3.68, 5.92, 8.07
-----------------------------------------------------------------------
Benchmark             Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------
BM_BytesCtor       6859 ns         6858 ns      6262143 bytes_per_second=217.495M/s
BM_BytesGet        5093 ns         5093 ns      8473049 bytes_per_second=390.99M/s
```


## After

```
Run on (96 X 2000.16 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x48)
  L1 Instruction 32 KiB (x48)
  L2 Unified 1024 KiB (x48)
  L3 Unified 39424 KiB (x2)
Load Average: 4.29, 7.98, 9.09
-----------------------------------------------------------------------
Benchmark             Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------
BM_BytesCtor       6596 ns         6595 ns     12388254 bytes_per_second=226.157M/s
BM_BytesGet        5075 ns         5074 ns     16639928 bytes_per_second=392.414M/s
```